+++ src/core/grid/Col.tsx
@@ -15,6 +15,7 @@ interface ColProps {
     applyFor?: Partial<Record<ColSizeType, number | undefined>>
     applyFrom?: [ColSizeType, number | undefined]
     ascendingApply?: boolean
+    className?: string
 }
 
 const Col: React.FC<ColProps> = ({