+++ src/form/form-control/FormControl.tsx
@@ -5,13 +5,20 @@ import styles from './form-control.module.scss'
 interface FormControlProps extends React.HTMLAttributes<HTMLDivElement> {
     children?: React.ReactNode
     error?: string | null
+    hideRequired?: boolean
 }
 
-const FormControl: React.FC<FormControlProps> = ({ children, className, error, ...props }) => {
+const FormControl: React.FC<FormControlProps> = ({
+    children,
+    className,
+    error,
+    hideRequired,
+    ...props
+}) => {
     return (
         <div className={cs(styles.root, className)} {...props}>
             {children}
-            {Boolean(error) && <p className={styles.error}>{error}</p>}
+            {!hideRequired && <>{Boolean(error) && <p className={styles.error}>{error}</p>}</>}
         </div>
     )
 }