+++ src/index.ts
@@ -6,7 +6,7 @@ export * from './hooks'
 export * from './form'
 export * from './core'
 export * from './misc'
-export * from './misc/icon/icons-types'
+export * from './types/icons-types'
 export * from './utils/icon-registry'
 
 export { InputBaseProps, ActionItemVariantsType }