+++ src/misc/icon/icons-types.ts
@@ -1,3 +1,0 @@
-import { icons } from '../../utils/available-icons'
-
-export type IconType = (typeof icons)[number]