+++ src/types/globals.d.ts
@@ -1,16 +1,5 @@
 declare global {
     type IconNameType = any
-    type ActionItemVariantsType =
-        | 'default'
-        | 'primary'
-        | 'secondary'
-        | 'success'
-        | 'info'
-        | 'danger'
-        | 'warning'
-        | 'light'
-        | 'dark'
-        | 'flat'
 }
 
 export {}