+++ src/types/icons-types.ts
@@ -1,0 +1,3 @@
+import { icons } from '../utils/available-icons'
+
+export type IconType = (typeof icons)[number]